backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==4.0.7
django-cors-headers==3.13.0
djangorestframework==3.13.1
flake8==5.0.4
idna==3.3
mccabe==0.7.0
mssql-django==1.1.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pyodbc==4.0.34
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.12
